chg: refactored classes to be in separate file
