+++ src/view/show_mode/editor/filter_settings_item.py
@@ -22,12 +22,12 @@ from PySide6.QtWidgets import (
 
 from model.filter import Filter, FilterTypeEnumeration
 from utility import resource_path
+from view.show_mode.editor.node_editor_widgets.chaser_editor.color_chaser_widget import ColorChaserFilterConfigWidget
 from view.show_mode.editor.node_editor_widgets.cue_editor import CueEditor
 from view.show_mode.editor.node_editor_widgets.pan_tilt_constant.pan_tilt_constant_widget import PanTiltConstantWidget
 from view.show_mode.effect_stacks.filter_config_widget import EffectsStackFilterConfigWidget
 
 from .node_editor_widgets.autotracker_settings import AutotrackerSettingsWidget
-from .node_editor_widgets.color_chaser_widget import ColorChaserFilterConfigWidget
 from .node_editor_widgets.color_mixing_setup_widget import ColorMixingSetupWidget
 from .node_editor_widgets.color_to_colorwheel_adapter_settings_widget import ColorToColorwheelAdapterSetupWidget
 from .node_editor_widgets.column_select import ColumnSelect
